Combine variable-length and loc-clust tests in 1 script
